test/requirements.txt
requests==2.10.0
Pillow==3.4.2
